Moved annotation methods into FlatFile.java so that they are inherited by all config classes
